backend/requirements.txt
Django>=4.0,<5.0
djangorestframework
requests
django-cors-headers
python-dotenv
